+++ src/isal/igzip.py
@@ -22,6 +22,7 @@
 Library to speed up its methods."""
 
 import argparse
+import functools
 import gzip
 import io
 import os
@@ -36,6 +37,8 @@ _COMPRESS_LEVEL_FAST = isal_zlib.ISAL_BEST_SPEED
 _COMPRESS_LEVEL_TRADEOFF = isal_zlib.ISAL_DEFAULT_COMPRESSION
 _COMPRESS_LEVEL_BEST = isal_zlib.ISAL_BEST_COMPRESSION
 _BLOCK_SIZE = 64*1024
+
+BUFFER_SIZE = _compression.BUFFER_SIZE
 
 
 # The open method was copied from the python source with minor adjustments.
@@ -145,8 +148,56 @@ class IGzipFile(gzip.GzipFile):
 # to do so in pure python.
 class _IGzipReader(_compression.DecompressReader):
     def __init__(self, fp):
-        super().__init__(fp, isal_zlib.decompressobj,
+        super().__init__(gzip._PaddedFile(fp), isal_zlib.decompressobj,
+                         trailing_error=isal_zlib.IsalError,
                          wbits=16 + isal_zlib.MAX_WBITS)
+
+    # Created by mixing and matching gzip._GzipReader and
+    # _compression.DecompressReader
+    def read(self, size=-1):
+        if size < 0:
+            return self.readall()
+        # size=0 is special because decompress(max_length=0) is not supported
+        if not size:
+            return b""
+
+        # For certain input data, a single
+        # call to decompress() may not return
+        # any data. In this case, retry until we get some data or reach EOF.
+        uncompress = b""
+        while True:
+            if self._decompressor.eof:
+                buf = (self._decompressor.unused_data or
+                       self._fp.read(BUFFER_SIZE))
+                if not buf:
+                    break
+                # Continue to next stream.
+                self._decompressor = self._decomp_factory(
+                    **self._decomp_args)
+                try:
+                    uncompress = self._decompressor.decompress(buf, size)
+                except self._trailing_error:
+                    # Trailing data isn't a valid compressed stream; ignore it.
+                    break
+            else:
+                # Read a chunk of data from the file
+                buf = self._fp.read(BUFFER_SIZE)
+                uncompress = self._decompressor.decompress(buf, size)
+            if self._decompressor.unconsumed_tail != b"":
+                self._fp.prepend(self._decompressor.unconsumed_tail)
+            elif self._decompressor.unused_data != b"":
+                # Prepend the already read bytes to the fileobj so they can
+                # be seen by _read_eof() and _read_gzip_header()
+                self._fp.prepend(self._decompressor.unused_data)
+
+            if uncompress != b"":
+                break
+            if buf == b"":
+                raise EOFError("Compressed file ended before the "
+                               "end-of-stream marker was reached")
+
+        self._pos += len(uncompress)
+        return uncompress
 
 
 # Plagiarized from gzip.py from python's stdlib.
@@ -161,6 +212,7 @@ def compress(data, compresslevel=_COMPRESS_LEVEL_BEST, *, mtime=None):
     return buf.getvalue()
 
 
+# Unlike stdlib, do not use the roundabout way of doing this via a file.
 def decompress(data):
     """Decompress a gzip compressed string in one shot.
     Return the decompressed string.
@@ -174,44 +226,49 @@ def main():
         "A simple command line interface for the igzip module. "
         "Acts like igzip.")
     parser.add_argument("file")
-    parser.add_argument("--fast", action="store_true",
-                        help="use fastest compression")
-    parser.add_argument("--best", action="store_true",
-                        help="use best compression")
-    parser.add_argument("-d", "--decompress", action="store_false",
-                        dest="compress",
-                        help="Decompress the file instead of compressing.")
+    compress_group = parser.add_mutually_exclusive_group()
+    compress_group.add_argument(
+        "-0", "--fast", action="store_const", dest="compresslevel",
+        const=_COMPRESS_LEVEL_FAST,
+        help="use compression level 0 (fastest)")
+    compress_group.add_argument(
+        "-1", action="store_const", dest="compresslevel",
+        const=1,
+        help="use compression level 1")
+    compress_group.add_argument(
+        "-2", action="store_const", dest="compresslevel",
+        const=2,
+        help="use compression level 2 (default)")
+    compress_group.add_argument(
+        "-3", "--best", action="store_const", dest="compresslevel",
+        const=_COMPRESS_LEVEL_BEST,
+        help="use compression level 3 (best)")
+    compress_group.add_argument(
+        "-d", "--decompress", action="store_false",
+        dest="compress",
+        help="Decompress the file instead of compressing.")
     args = parser.parse_args()
 
-    if args.fast:
-        compresslevel = _COMPRESS_LEVEL_FAST
-    elif args.best:
-        compresslevel = _COMPRESS_LEVEL_BEST
-    else:
-        compresslevel = _COMPRESS_LEVEL_TRADEOFF
+    compresslevel = args.compresslevel or _COMPRESS_LEVEL_TRADEOFF
 
     if args.compress:
         out_filename = args.file + ".gz"
-        with io.open(args.file, "rb") as in_file:
-            with open(out_filename, mode="rb", compresslevel=compresslevel
-                      ) as out_file:
-                while True:
-                    block = in_file.read(_BLOCK_SIZE)
-                    if block == b"":
-                        break
-                    out_file.write(block)
+        out_open = functools.partial(open, compresslevel=compresslevel)
+        in_open = io.open
     else:
         base, extension = os.path.splitext(args.file)
         if extension != ".gz":
             raise ValueError("Can only decompress files with a .gz extension")
         out_filename = base
-        with open(args.file, "rb") as in_file:
-            with io.open(out_filename, mode="rb") as out_file:
-                while True:
-                    block = in_file.read(_BLOCK_SIZE)
-                    if block == b"":
-                        break
-                    out_file.write(block)
+        out_open = io.open
+        in_open = open
+    with in_open(args.file, "rb") as in_file:
+        with out_open(out_filename, "wb") as out_file:
+            while True:
+                block = in_file.read(_BLOCK_SIZE)
+                if block == b"":
+                    break
+                out_file.write(block)
 
 
 if __name__ == "__main__":